Added comments section
Added a comments section that creates comments and update state when a new one is added.

Created a custom hook to fetchSinglePhoto from api for easier reusability.

Added a util function to format the date of the comments.

Changed decodeToken function to return the whole token instead of just the userId for better scaleability.