Updated to a cloud cluster instead of local
Updated mongodb to a cloud cluster to persist data across multiple devices.

Added change password functionality with input clearing and a success/error message.

Added get user functionality on the server.
Added a custom hook that fetches the current user info rom the auth token for better reusability.

Fixed email state not updating when email is changed on the server.