Added new functionality, made changes and improvements
Fixed photos showing scaled in photo details page, by creating a responsive photo component that changes based on photo aspect ratio.

Improved home page layout by changing css.

Added a feature to change user description in user profile page and removed skills sections as it was not needed.

Added a user description POST action on backend to handle description updates.

Changed useFetchSingle photo a bit.

Added a like photo button in an overlay on images and implemented like functionality which updates the UI on like and when like is removed.