Added missing edit post functionality
Improved folder structure for better viewing experience.

Added edit post CRUD operations to backend and implemented in the front end using an edit modal.

Fixed minor bugs and improved some animations.

Fixed error handling on email change input.

Made uploadModal reusable and used it on profile picture upload as well.

Added a title to all record in the DB using a script to populate them with a title that matches the current caption to ensure consistency when creating the edit post option.